[FIX] stock_quantity_history_location: qty by location
When a qty at date is asked the location context could be overriden if
the company_owned one was inserted after we removed it. We ensure to
remove it to always obtain the right values.

As a side effect, stock_account valuation is now correctly computed when
AVCO or Standard methods are implied. So should be with FIFO manual (as
it computes the quantities the same way) but doubtful in FIFO real time.