[REF]stock_quant_history: Invert snapshot logic to use allowed locations
The previous implementation of the stock quant history snapshot used a
negative logic, ignoring a list of specified location usages. This was
causing issues with virtual production locations and made the module
hard to maintain.

This commit refactors the logic to be positive. Instead of ignoring
locations, the logic now only includes locations with the 'internal'
usage type. This simplifies the code, improves maintainability, and
resolves the issue with virtual locations.

The `_ignored_location_usage` method has been renamed to
`_allowed_location_usage` to better reflect its new purpose. The
corresponding unit tests have also been updated to align with the new
logic.