[FIX] stock_picking_report_valued: wrong prices
The way the fields were calculated has been modified, since after performing the new() function, it recalculates the fields and loses them, only showing the unit price at the end. Therefore, it has been modified so that if it has a quantity different from the sales order or a different UOM, it first calculates the subtotal price and then performs the new() function. This way, the prices are calculated correctly in the report and the lines.